# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Norv_ai – Smart IT Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/setup:
    post:
      operationId: completeSetup
      tags: [profile]
      summary: Complete 6-step onboarding wizard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /profile:
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /profile/change-password:
    post:
      operationId: changePassword
      tags: [profile]
      summary: Change account password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
  /lectures:
    get:
      operationId: getLectures
      tags: [lectures]
      summary: List lectures with optional filters
      parameters:
        - in: query
          name: specialization
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
        - in: query
          name: curriculumId
          schema:
            type: number
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: saved
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"

    post:
      operationId: createLecture
      tags: [lectures]
      summary: Create a new lecture (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
  /lectures/{id}:
    get:
      operationId: getLecture
      tags: [lectures]
      summary: Get a single lecture
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateLecture
      tags: [lectures]
      summary: Update lecture (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"

    delete:
      operationId: deleteLecture
      tags: [lectures]
      summary: Delete lecture (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /lectures/{id}/save:
    post:
      operationId: saveLecture
      tags: [lectures]
      summary: Save lecture to favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

    delete:
      operationId: unsaveLecture
      tags: [lectures]
      summary: Remove lecture from favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /lectures/{id}/complete:
    post:
      operationId: completeLecture
      tags: [lectures]
      summary: Mark lecture as completed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Curricula ────────────────────────────────────────────────────────────
...
  /summaries:
    get:
      operationId: getSummaries
      tags: [summary]
      summary: List user's summaries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Summary"

    post:
      operationId: createSummary
      tags: [summary]
      summary: Generate smart summary from text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryInput"
      responses:
        "201":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
  /summaries/{id}:
    get:
      operationId: getSummary
      tags: [summary]
      summary: Get a single summary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"

    delete:
      operationId: deleteSummary
      tags: [summary]
      summary: Delete a summary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── Code Debugger ─────────────────────────────────────────────────────────
...
  /subscriptions/request:
    post:
      operationId: requestSubscription
      tags: [subscriptions]
      summary: Submit subscription request with CliQ receipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequestInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionRequest"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # Auth
...
    User:
      type: object
      required: [id, username, fullName, role, setupComplete, createdAt]
      properties:
        id:
          type: number
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [student, admin]
        setupComplete:
          type: boolean
        university:
          type: ["string", "null"]
        major:
          type: ["string", "null"]
        yearOfStudy:
          type: ["number", "null"]
        specialization:
          type: ["string", "null"]
        skillLevel:
          type: ["string", "null"]
        knownLanguages:
          type: array
          items:
            type: string
        subscriptionActive:
          type: boolean
        subscriptionExpiry:
          type: ["string", "null"]
        createdAt:
          type: string

    # Profile / Onboarding
    SetupInput:
      type: object
      required: [specialization, skillLevel]
      properties:
        university:
          type: string
        major:
          type: string
        yearOfStudy:
          type: number
        specialization:
          type: string
        skillAnswers:
          type: array
          items:
            type: string
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        knownLanguages:
          type: array
          items:
            type: string
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        university:
          type: string
        major:
          type: string
        yearOfStudy:
          type: number
        specialization:
          type: string
        skillLevel:
          type: string
        knownLanguages:
          type: array
          items:
            type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # Dashboard
...
    Lecture:
      type: object
      required: [id, title, youtubeId, instructor, specialization, difficulty, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        youtubeId:
          type: string
        youtubeUrl:
          type: string
        instructor:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        specialization:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        durationMinutes:
          type: ["number", "null"]
        curriculumId:
          type: ["number", "null"]
        curriculumName:
          type: ["string", "null"]
        isSaved:
          type: boolean
        isCompleted:
          type: boolean
        createdAt:
          type: string
    LectureInput:
      type: object
      required: [title, youtubeUrl, specialization, difficulty]
      properties:
        title:
          type: string
        youtubeUrl:
          type: string
        instructor:
          type: string
        description:
          type: string
        specialization:
          type: string
        difficulty:
          type: string
        durationMinutes:
          type: number
        curriculumId:
          type: number
    LectureUpdate:
      type: object
      properties:
        title:
          type: string
        youtubeUrl:
          type: string
        instructor:
          type: string
        description:
          type: string
        specialization:
          type: string
        difficulty:
          type: string
        durationMinutes:
          type: number
        curriculumId:
          type: ["number", "null"]

    # Curricula
...
    Summary:
      type: object
      required: [id, title, topic, summary, keyPoints, createdAt]
      properties:
        id:
          type: number
        title:
          type: ["string", "null"]
        topic:
          type: string
        originalText:
          type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        technicalTerms:
          type: array
          items:
            type: string
        difficultyLevel:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    SummaryInput:
      type: object
      required: [text, topic]
      properties:
        text:
          type: string
          minLength: 50
        topic:
          type: string
        title:
          type: string

    # Code Debugger
...
    SubscriptionRequestInput:
      type: object
      required: [plan, receiptUrl]
      properties:
        plan:
          type: string
          enum: ["3months", "6months", "1year"]
        receiptUrl:
          type: string
    SubscriptionRequest:
      type: object
      required: [id, userId, plan, status, receiptUrl, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        username:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        plan:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        receiptUrl:
          type: string
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...