# lib/api-zod/src/generated/types/subscriptionRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Norv_ai – Smart IT Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionRequestInputPlan } from './subscriptionRequestInputPlan';

export interface SubscriptionRequestInput {
  plan: SubscriptionRequestInputPlan;
  receiptUrl: string;
}